refactor link to be more reusable, remove forwardRef util
