+++ frontend/src/app/UIkit/Link/Link/index.ts
@@ -1,0 +1,1 @@
+export * from './Link'
