+++ frontend/src/app/UIkit/Link/NavLink/NavLink.ts
@@ -1,0 +1,3 @@
+import { linkStyled } from '../Link.styles'
+
+const NavLinkStyled = linkStyled('a')
