+++ frontend/src/app/UIkit/Link/NavLink/index.ts
@@ -1,2 +1,1 @@
-export * from './Link'
 export * from './NavLink'
