--- frontend/src/app/utils/core/forwardRef.ts
@@ -1,60 +1,0 @@
-/**
- * Part of this code is taken from https://github.com/nextui-org/nextui/blob/main/packages/react/src/utils/system.ts
- */
-
-import React, { forwardRef as baseForwardRef } from 'react'
-import { CSS } from '@stitches/react'
-
-export type As<Props = any> = React.ElementType<Props>
-
-export type OmitCommonProps<Target, OmitAdditionalProps extends keyof any = never> = Omit<Target,
-'transition' | 'as' | 'color' | OmitAdditionalProps>
-
-export type RightJoinProps<SourceProps extends object = {},
-  OverrideProps extends object = {},
-  > = OmitCommonProps<SourceProps, keyof OverrideProps> & OverrideProps
-
-export type MergeWithAs<ComponentProps extends object,
-  AsProps extends object,
-  AdditionalProps extends object = {},
-  AsComponent extends As = As,
-  > = RightJoinProps<ComponentProps, AdditionalProps> &
-  RightJoinProps<AsProps, AdditionalProps> & {
-    as?: AsComponent
-  }
-
-export interface ComponentWithAs<Component extends As, Props extends object = {}> {
-  <AsComponent extends As = Component>(
-    props: MergeWithAs<React.ComponentProps<Component>,
-    React.ComponentProps<AsComponent>,
-    Props,
-    AsComponent>,
-  ): JSX.Element
-
-  displayName?: string
-  propTypes?: React.WeakValidationMap<any>
-  contextTypes?: React.ValidationMap<any>
-  defaultProps?: Partial<any>
-  id?: string
-}
-
-/**
- * Extract the props of a React element or component
- */
-export type PropsOf<T extends As> = React.ComponentPropsWithoutRef<T> & {
-  as?: As
-}
-
-export type HTMLNextUIProps<T extends As> = Omit<PropsOf<T>, 'ref'> & {
-  as?: As
-  css?: CSS
-}
-
-export function forwardRef<Props extends object, Component extends As> (
-  component: React.ForwardRefRenderFunction<any,
-  RightJoinProps<PropsOf<Component>, Props> & {
-    as?: As
-  }>
-) {
-  return baseForwardRef(component) as unknown as ComponentWithAs<Component, Props>
-}
